GTiff: fix trunk potential int32 overflow in HasOnlyNoDataT() that could cause crashes in tiff_read_irregular_tile_size_jpeg_in_tiff test
git-svn-id: https://svn.osgeo.org/gdal/trunk@34615 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/gtiff/geotiff.cpp
@@ -7209,10 +7209,10 @@ bool GTiffDataset::HasOnlyNoDataT( const T* pBuffer, int nWidth, int nHeight,
     // Fast test: check the 4 corners and the middle pixel
     for(int iBand = 0; iBand < nComponents; iBand++ )
     {
-        if( !(IsEqualToNoData(pBuffer[(nWidth-1) * nComponents + iBand], noDataValue) &&
-              IsEqualToNoData(pBuffer[(nHeight-1) * nLineStride * nComponents + iBand], noDataValue) &&
-              IsEqualToNoData(pBuffer[((nHeight-1) * nLineStride + nWidth - 1) * nComponents + iBand], noDataValue) &&
-              IsEqualToNoData(pBuffer[((nHeight-1)/2 * nLineStride + (nWidth - 1)/2) * nComponents + iBand], noDataValue)) )
+        if( !(IsEqualToNoData(pBuffer[(size_t)(nWidth-1) * nComponents + iBand], noDataValue) &&
+              IsEqualToNoData(pBuffer[(size_t)(nHeight-1) * nLineStride * nComponents + iBand], noDataValue) &&
+              IsEqualToNoData(pBuffer[((size_t)(nHeight-1) * nLineStride + nWidth - 1) * nComponents + iBand], noDataValue) &&
+              IsEqualToNoData(pBuffer[((size_t)(nHeight-1)/2 * nLineStride + (nWidth - 1)/2) * nComponents + iBand], noDataValue)) )
         {
             return false;
         }
@@ -7223,7 +7223,7 @@ bool GTiffDataset::HasOnlyNoDataT( const T* pBuffer, int nWidth, int nHeight,
     {
         for(int iX = 0; iX < nWidth * nComponents; iX++ )
         {
-            if( !IsEqualToNoData(pBuffer[iY * nLineStride * nComponents + iX], noDataValue) )
+            if( !IsEqualToNoData(pBuffer[iY * (size_t)nLineStride * nComponents + iX], noDataValue) )
             {
                 return false;
             }
